rag_pipeline/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
psycopg2-binary==2.9.9
pika==1.3.2
redis==5.0.3
httpx==0.27.0
python-multipart==0.0.9
sentence-transformers==2.5.1
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
pydantic==2.6.3
python-dotenv==1.0.1
